+++ ixtb-nextjs/src/lib/endpoints/external/members/deleteMembersEndpoint.ts
@@ -1,5 +1,10 @@
-import { deleteMembersSchema, kByTypes } from "fimidx-core/definitions/index";
+import {
+  deleteMembersSchema,
+  kByTypes,
+  kFimidxPermissions,
+} from "fimidx-core/definitions/index";
 import { deleteMembers } from "fimidx-core/serverHelpers/index";
+import { checkPermissionGroupThenProjectThenOrg } from "../../../serverHelpers/permissions";
 import { NextClientTokenAuthenticatedEndpointFn } from "../../types";
 import { sanitizeDeleteMembersInput } from "../../utils/sanitizeKId0.js";
 
@@ -13,6 +18,14 @@ export const deleteMemberEndpoint: NextClientTokenAuthenticatedEndpointFn<
 
   const input = deleteMembersSchema.parse(await req.json());
   sanitizeDeleteMembersInput(input);
+
+  await checkPermissionGroupThenProjectThenOrg({
+    clientToken,
+    groupId: input.query.groupId,
+    projectId: input.query.projectId,
+    action: kFimidxPermissions.member.remove,
+  });
+
   await deleteMembers({
     by: clientToken.id,
     byType: kByTypes.clientToken,
