+++ ixtb-nextjs/src/lib/endpoints/internal/orgs/deleteOrgEndpoint.ts
@@ -1,7 +1,8 @@
 import { deleteOrgSchema } from "@/src/definitions/org";
-import { kByTypes } from "fimidx-core/definitions/index";
+import { kByTypes, kFimidxPermissions } from "fimidx-core/definitions/index";
 import { kId0 } from "fimidx-core/definitions/system";
 import { deleteGroups } from "fimidx-core/serverHelpers/index";
+import { requirePermissionForUser } from "../../../serverHelpers/permissions";
 import { NextUserAuthenticatedEndpointFn } from "../../types";
 import { sanitizeDeleteOrgInput } from "../../utils/sanitizeKId0.js";
 
@@ -18,6 +19,13 @@ export const deleteOrgEndpoint: NextUserAuthenticatedEndpointFn<void> = async (
     id: pathParams.orgId,
   });
   sanitizeDeleteOrgInput(input);
+
+  await requirePermissionForUser({
+    userId,
+    orgId: input.id,
+    action: kFimidxPermissions.group.delete,
+    target: input.id,
+  });
 
   await deleteGroups({
     query: {
